Fix warnings found by GCC
My previous patch addressed all LLVM warnings, this one addresses all the GCC ones as well (mostly signed / unsigned mix).

The patch also turns on -Wall -Werror.

+++ src/wasm-interpreter.h
@@ -248,7 +248,7 @@ private:
         NOTE_EVAL1(flow.value);
         int64_t index = flow.value.getInteger();
         Name target = curr->default_;
-        if (index >= 0 && index < curr->targets.size()) {
+        if (index >= 0 && (size_t)index < curr->targets.size()) {
           target = curr->targets[index];
         }
         // This is obviously very inefficient. This should be a cached data structure
